added dummy data and also created post route for username and password to return all user data
