+++ controllers/UserController.js
@@ -1,5 +1,5 @@
 const UserModel = require('../models/UserModel');
-
+////////////////////////////////////////////////////////////////////////
 async function getAllUsers(req, res) {
     try {
         const users = await UserModel.getAllUsers();
@@ -9,9 +9,27 @@ async function getAllUsers(req, res) {
         res.status(500).json({ error: 'Error getting users' });
     }
 }
+////////////////////////////////////////////////////////////////////////
+async function loginUser(req, res) {
+    const { username, password } = req.body;
+    try {
+        const user = await UserModel.loginUser(username, password);
 
+        if (!user) {
+            return res.status(401).json({ error: 'Invalid credentials' });
+        }
+
+        return res.json(user);
+
+    } catch (error) {
+        console.error('Error logging in:', error);
+        return res.status(500).json({ error: 'Error logging in' });
+    }
+}
+////////////////////////////////////////////////////////////////////////
 
 module.exports = {
     getAllUsers,
+    loginUser,
     // Add more controller methods
 };
