+++ data/dummyUserData.js
@@ -1,0 +1,34 @@
+const dummyUserData = [
+    {
+        username: 'user1',
+        password: 'pw1',
+        fullName: 'John Doe',
+        address1: '123 Main Street',
+        address2: 'Apt 456',
+        city: 'Los Angeles',
+        state: 'CA',
+        zipcode: '90001',
+    },
+    {
+        username: 'user2',
+        password: 'pw2',
+        fullName: 'Alice Smith',
+        address1: '456 Elm Street',
+        address2: '',
+        city: 'New York',
+        state: 'NY',
+        zipcode: '10001',
+    },
+    {
+        username: 'user3',
+        password: 'pw3',
+        fullName: 'Bob Johnson',
+        address1: '789 Oak Street',
+        address2: 'Suite 101',
+        city: 'Chicago',
+        state: 'IL',
+        zipcode: '60601',
+    },
+];
+
+module.exports = dummyUserData;
