+++ models/UserModel.js
@@ -1,54 +1,37 @@
 //const db = require('../db');
-
-// Function to get all users from the database
+const dummyUserData = require('../data/dummyUserData');
+////////////////////////////////////////////////////////////////////////
 async function getAllUsers() {
-    const query = 'SELECT * FROM users';
+    //const query = 'SELECT * FROM users';
     try {
         //const [results] = await db.query(query);
+        //return results;
 
         //FIXME Temporary Dummy Data
-        const dummyUserData = [
-            {
-                username: 'user1',
-                password: 'pw1',
-                fullName: 'John Doe',
-                address1: '123 Main Street',
-                address2: 'Apt 456',
-                city: 'Los Angeles',
-                state: 'CA',
-                zipcode: '90001',
-            },
-            {
-                username: 'user2',
-                password: 'pw2',
-                fullName: 'Alice Smith',
-                address1: '456 Elm Street',
-                address2: '',
-                city: 'New York',
-                state: 'NY',
-                zipcode: '10001',
-            },
-            {
-                username: 'user3',
-                password: 'pw3',
-                fullName: 'Bob Johnson',
-                address1: '789 Oak Street',
-                address2: 'Suite 101',
-                city: 'Chicago',
-                state: 'IL',
-                zipcode: '60601',
-            },
-        ];
-
-        return dummyUserData;
-
-
+        const results = dummyUserData;
+        return results;
     } catch (error) {
         throw error;
     }
 }
+////////////////////////////////////////////////////////////////////////
+async function loginUser(username, password) {
+    //const query = 'SELECT * FROM users WHERE username = ? AND password = ?';
+    try {
+        //const [results] = await db.query(query, [username, password]);
+        //return results.length === 0 ? null : results[0];
+        //FIXME Temporary Dummy Data
+        const user = dummyUserData.find((user) => user.username === username && user.password === password);
+        return user;
 
+    } catch (error) {
+        //console.error('Error fetching user by credentials:', error);
+        throw error;
+    }
+}
+////////////////////////////////////////////////////////////////////////
 module.exports = {
     getAllUsers,
+    loginUser,
     // Add more functions for other user-related database operations
 };
