+++ routes/User.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const controller = require('../controllers/UserController');
 
 router.get('/', controller.getAllUsers);
+router.post('/login', controller.loginUser);
 
 
 module.exports = router;
